Add trait ReplicaManager and ReplicaManager{Zk,Raft}
