+++ core/src/test/scala/unit/kafka/server/epoch/OffsetsForLeaderEpochTest.scala
@@ -42,7 +42,7 @@ class OffsetsForLeaderEpochTest {
   private val metrics = new Metrics
   private val alterIsrManager = TestUtils.createAlterIsrManager()
   private val tp = new TopicPartition("topic", 1)
-  private var replicaManager: ReplicaManager = _
+  private var replicaManager: ReplicaManagerZk = _
   private var quotaManager: QuotaManagers = _
 
   @Before
@@ -65,7 +65,7 @@ class OffsetsForLeaderEpochTest {
     replay(mockLog, logManager)
 
     // create a replica manager with 1 partition that has 1 replica
-    replicaManager = new ReplicaManager(config, metrics, time, null, null, logManager, new AtomicBoolean(false),
+    replicaManager = new ReplicaManagerZk(config, metrics, time, null, null, logManager, new AtomicBoolean(false),
       quotaManager, new BrokerTopicStats,
       new MetadataCache(config.brokerId), new LogDirFailureChannel(config.logDirs.size), alterIsrManager)
     val partition = replicaManager.createPartition(tp)
@@ -88,7 +88,7 @@ class OffsetsForLeaderEpochTest {
     replay(logManager)
 
     //create a replica manager with 1 partition that has 0 replica
-    replicaManager = new ReplicaManager(config, metrics, time, null, null, logManager, new AtomicBoolean(false),
+    replicaManager = new ReplicaManagerZk(config, metrics, time, null, null, logManager, new AtomicBoolean(false),
       quotaManager, new BrokerTopicStats,
       new MetadataCache(config.brokerId), new LogDirFailureChannel(config.logDirs.size), alterIsrManager)
     replicaManager.createPartition(tp)
@@ -113,7 +113,7 @@ class OffsetsForLeaderEpochTest {
     replay(logManager)
 
     //create a replica manager with 0 partition
-    replicaManager = new ReplicaManager(config, metrics, time, null, null, logManager, new AtomicBoolean(false),
+    replicaManager = new ReplicaManagerZk(config, metrics, time, null, null, logManager, new AtomicBoolean(false),
       quotaManager, new BrokerTopicStats,
       new MetadataCache(config.brokerId), new LogDirFailureChannel(config.logDirs.size), alterIsrManager)
